requirements.txt
future==0.18.2
ipython-genutils==0.2.0
numpy==1.20.2
parso==0.8.2
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyglet==1.5.0
Pygments==2.8.1
pymdptoolbox==4.0b3
pyparsing==2.4.7
scipy==1.6.2
traitlets==5.0.5
wcwidth==0.2.5
